Deleting and inserting columns to align all the data needed for processing into three columns.
Our titration method is two staged, first, titrating to an end point pH using whatever volume of titrant is necessary. The second stage involves adding fixed volumes of titrant and allowing the pH electrode to settle to equivalence points of pH.

During these titrations we collect data on: Volume of Titrant added in milliters (mL), the potential seen by the pH electrode in millivolts (mV) and the temperature seen in the sample via a temperature probe in degrees Celsius.