Wrapping up pH processing that was meant to finish on May 7th 2021.
Updated the names of pointers in the script to set the working directory to the right place, pulled updated file names with the new files, then altered the filename to run the rest of the script.

Altering the filename variable happens after clearing the environment, then running the script all the way through to process the data.