I am finishing the total alkalinity data process intended to be completed on May 7th 2021.
I pulled the data out for samples that were ran on 4/7/2021. I then put these pulled data files into their own directory so when I go to process they don't conflict with a different day.

I had to adjust the name for the bay standard in the samples only file for 4/7/2021 so that they would match what I entered in the metadata.