Uploaded the most recent data from the last alkalinity standard run. As well as the training session with MJ.
I double checked to see if the same min/max issue was happening in the alkalinity data file and I found no issues.

Most of the data looked pretty good. The first replicate of the Dickson standard was a bit off so I decided to exclude that and use the second two replicates.