Wow that toook waaaay longer than I wanted to.
I got most of the column names to comply with each other whether the data comes from archived .csv files from EcoWatch Lite or from the real time text file output and using the header provided to Lynn Dewitt for the real time data.

I had to force the real time header to drop "_cm" from the conductivity columns and change "psia" to "psir" since that is actually what we are measuring.

"psia" and "psir" are absolute and relative pressure. According to the manual for YSI sondes psia is only used when you have a specially vented sonde that can account for changes in pressure from the atmosphere, which we do not have.

I hate hate hate how this solution came together but I am out of time and need to leave it at that.